Cisol,
The client would like a space between the sub header on the landing page and the beginning of the text. This is something she should be able to do in WordPress, yes?

On the mobile she would like the same (above) carriage return between the sub header and the content .

Thank you!