1)      Menus should be just the main headers, not that crazy long list.  Question of course is how do you get to the sub menus easily?  Perhaps clicking on them opens a list of just that section?  Can that list be white and the main headers be gray so it doesn’t all look the same?  Some of the longer names wrap odd (ie see Probate, Estate & Trust Administration) This is a significant request. See note on #4
6)      When I’m in the page, ie About Rendigs, the only way to navigate through that area is to go back to the main menu and the long list…. Is there a way to display the sub-menus in that “category” for easy navigation?
Of course theres a way, but not within a reasonable time and budget. If you’d like to invest in optimizing the responsive experience, you can. We have provided basic responsive functionality per the contract. As a developer I was given ZERO design input from Sunrise or Rendigs on responsive factors prior to development. I went with a common setup that I ahve uses for very many satisfied clients. That said, I have spent nearly 4 hours building the nav to work in a way I hope will be satisfactory. Additional changes will have to be addressed as new work orders.